requirements.txt
python-binance==1.0.19
websocket-client==1.6.4
pandas==2.1.4
numpy==1.26.2